+++ java/com/google/tca/adapters/endorsement/Claim.java
@@ -29,11 +29,15 @@ import java.util.Map;
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Claim {
-  public static final String PES_CLAIM_PUBLISHER_V1_TYPE =
+  public static final String PES_CLAIM_PUBLISHER_LEGACY_TYPE =
       "https://github.com/pcit/pes/docs/claims/v1/publisher.md";
+  public static final String PES_CLAIM_PUBLISHER_TYPE =
+      "https://github.com/private-compute-infra-toolkit/public-endorsement-service/blob/main/docs/claims/publisher.md";
   public static final String PUBLISHER_ID_KEY = "publisher_id";
-  public static final String WORKLOAD_CLAIM_TYPE =
+  public static final String WORKLOAD_CLAIM_LEGACY_TYPE =
       "https://github.com/pcit/pes/docs/claims/v1/workload.md";
+  public static final String WORKLOAD_CLAIM_TYPE =
+      "https://github.com/private-compute-infra-toolkit/public-endorsement-service/blob/main/docs/claims/workload.md";
   public static final String WORKLOAD_ID_KEY = "workload_id";
 
   private final String type;
@@ -47,9 +51,9 @@ public class Claim {
       throw new IllegalArgumentException("Claim type is null");
     }
 
-    if (type.equals(PES_CLAIM_PUBLISHER_V1_TYPE)) {
+    if (type.equals(PES_CLAIM_PUBLISHER_LEGACY_TYPE) || type.equals(PES_CLAIM_PUBLISHER_TYPE)) {
       validatePublisherClaim(annotations);
-    } else if (type.equals(WORKLOAD_CLAIM_TYPE)) {
+    } else if (type.equals(WORKLOAD_CLAIM_LEGACY_TYPE) || type.equals(WORKLOAD_CLAIM_TYPE)) {
       validateWorkloadClaim(annotations);
     }