+++ java/com/google/tca/server/AwsInstanceMetadata.java
@@ -16,6 +16,27 @@
 
 package com.google.tca.server;
 
+import com.google.auto.value.AutoBuilder;
+
 /** Immutable carrier for IMDS data. */
 public record AwsInstanceMetadata(
-    String region, String accountId, String environment, String domain) {}
+    String region, String accountId, String environment, String domain) {
+
+  public static Builder builder() {
+    return new AutoBuilder_AwsInstanceMetadata_Builder();
+  }
+
+  /** Builder for {@link AwsInstanceMetadata}. */
+  @AutoBuilder
+  public interface Builder {
+    Builder setRegion(String region);
+
+    Builder setAccountId(String accountId);
+
+    Builder setEnvironment(String environment);
+
+    Builder setDomain(String domain);
+
+    AwsInstanceMetadata build();
+  }
+}