+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.3.0 (2026-06-03)
+
+
+### Features
+
+* Dynamic OIDC audience binding validation
+* feat: Add support for new claims .md location
+
 ## 0.2.0 (2026-05-29)